requirements.txt
discord.py==1.5.1
motor==2.3.0
umongo==2.3.0
